add new entity's attributes

+++ backend/src/main/java/com/entity/dto/ProductVariantDTO.java
@@ -8,5 +8,5 @@ public class ProductVariantDTO {
     private String color;
     private String imageUrl;
     private Integer stockQuantity;
-    private Double discount;
+    private Integer discount;
 }
